Refactor README and scripts: update paths, add new scripts, and implement prime number function
