+++ README.md
@@ -1,8 +1,8 @@
 ```console
-% uv run examples/cow.py
-% uv run examples/use-git-library.py
+% uv run scripts/cow.py
+% uv run scripts/use-git-library.py
 % uv run pytest
-% uv run mypy src/prime.py
+% uv run mypy src/hello_python_uve/prime.py
 ```
 
 The Python version can be decided by looking [here](https://devguide.python.org/versions/).
