+++ scripts/cow.py
